Moved fetch funcs to database API.
